requirements.txt
scrapy==1.5.2
PyMySQL
proto-plus
kafka-python
redis
bloom-filter
python-snappy
persist-queue
flask